+++ packages/ohm-js/src/deferredInit.js
@@ -1,0 +1,8 @@
+'use strict';
+
+// --------------------------------------------------------------------
+// Imports
+// --------------------------------------------------------------------
+
+require('./grammarDeferredInit');
+require('./semanticsDeferredInit');
