+++ packages/ohm-js/src/grammarDeferredInit.js
@@ -1,0 +1,13 @@
+'use strict';
+
+// --------------------------------------------------------------------
+// Imports
+// --------------------------------------------------------------------
+
+const Grammar = require('./Grammar');
+
+// --------------------------------------------------------------------
+// Private stuff
+// --------------------------------------------------------------------
+
+Grammar.BuiltInRules = require('../dist/built-in-rules');
